+++ src/main/java/com/mipt/hsse/hssetechbackend/payments/providers/tinkoff/entities/requests/CreatePaymentSessionTinkoffEntity.java
@@ -1,5 +1,6 @@
 package com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.entities.requests;
 
+import com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.signing.TinkoffSign;
 import lombok.Getter;
 
 @Getter
