+++ src/test/java/com/mipt/hsse/hssetechbackend/payments/providers/tinkoff/helpers/SignableTinkoffApiClientTest.java
@@ -1,7 +1,7 @@
 package com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.helpers;
 
 import com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.entities.requests.TinkoffRequestBase;
-import com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.entities.requests.TinkoffSign;
+import com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.signing.TinkoffSign;
 import com.mipt.hsse.hssetechbackend.payments.providers.tinkoff.signing.RequestsSignerBase;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
